Making some adjustments

+++ lib/globals.dart
@@ -4,5 +4,5 @@ class Global {
   static int roundId;
   static double roundBet;
   static int playerTurnId;
-  static double playerMoney;
+  static int playerMoney;
 }
